commands

+++ src/main/kotlin/org/quintilis/economy/commands/listing/ListingCommand.kt
@@ -12,6 +12,10 @@ import org.quintilis.economy.dao.ListingDao
 import org.quintilis.economy.dao.PlayerDao
 import org.quintilis.economy.entities.PlayerEntity
 import org.quintilis.economy.entities.listings.Listing
+import org.quintilis.economy.entities.listings.ListingStatus
+import org.quintilis.economy.entities.transactions.AdminTransaction
+import org.quintilis.economy.entities.transactions.Transaction
+import org.quintilis.economy.entities.transactions.TransactionType
 import org.quintilis.economy.managers.DatabaseManager
 
 class ListingCommand : BaseCommand(
@@ -63,9 +67,7 @@ class ListingCommand : BaseCommand(
                         Argument.component("item_name", item.displayName()),
                         Argument.numeric("quantity", listing.quantity),
                         Argument.numeric("price", listing.askingPricePerItem),
-                        Argument.component("status", Component.translatable(
-                            "listing.status.${listing.status.name}",
-                        ))
+                        Argument.component("status", listing.status.getComponent())
                     )
                 }!!
             }
@@ -79,6 +81,28 @@ class ListingCommand : BaseCommand(
         val listingId = args[0].toIntOrNull() ?: return this.argumentsMissing(sender)
 
         val player = sender as Player
+
+        val listing = listingDao.findById(listingId)
+
+        if(listing == null){
+            sender.sendMessage {
+                Component.translatable(
+                    "listing.remove.error.not_found",
+                    Argument.numeric("id", listingId)
+                )
+            }
+            return true
+        }
+        if(listing.status != ListingStatus.ACTIVE){
+            sender.sendMessage {
+                Component.translatable(
+                    "listing.remove.error.listing_no_active",
+                    Argument.component("state", listing.status.getComponent())
+
+                )
+            }
+            return true
+        }
 
         // 1. O DAO agora faz a checagem E a atualização de uma só vez
         val cancelledListing = listingDao.removeListingById(listingId, player.uniqueId)
@@ -122,7 +146,7 @@ class ListingCommand : BaseCommand(
             return true
         }
 
-        val quantity = args.getOrNull(2)?.toIntOrNull() ?: currentItem.amount
+        val quantity = args.getOrNull(1)?.toIntOrNull() ?: currentItem.amount
         if(quantity > currentItem.amount || quantity > 64){
             sender.sendMessage(
                 Component.translatable(
@@ -132,9 +156,12 @@ class ListingCommand : BaseCommand(
             return true
         }
 
-        val itemBytes = currentItem.serializeAsBytes()
+        val itemParaSalvar = currentItem.clone()
+        itemParaSalvar.amount = 1
 
-        println(sender.uniqueId.toString())
+        val itemBytes = itemParaSalvar.serializeAsBytes()
+
+//        println(sender.uniqueId.toString())
 
         val listing = Listing(
             sellerUuid = sender.uniqueId,
@@ -171,7 +198,15 @@ class ListingCommand : BaseCommand(
 
         val playerEntity = playerDao.findById(player.uniqueId) ?: return false
         playerEntity.points += points.toInt()
-
+        val savedTransaction = Transaction(
+            playerId = player.uniqueId,
+            transactionType = TransactionType.ADMIN_GIVE,
+            change = points.toInt()
+        ).save<Transaction>()
+        AdminTransaction(
+            transactionId = savedTransaction.id!!,
+            adminId = (sender as Player).uniqueId
+        ).save<AdminTransaction>()
         playerEntity.save<PlayerEntity>()
 
         sender.sendMessage {
@@ -195,6 +230,16 @@ class ListingCommand : BaseCommand(
 
         val playerEntity = playerDao.findById(player.uniqueId) ?: return false
         playerEntity.points -= args[1].toInt()
+
+        val savedTransaction = Transaction(
+            playerId = player.uniqueId,
+            transactionType = TransactionType.ADMIN_TAKE,
+            change = args[1].toInt()
+        ).save<Transaction>()
+        AdminTransaction(
+            transactionId = savedTransaction.id!!,
+            adminId = (sender as Player).uniqueId
+        ).save<AdminTransaction>()
 
         val savedEntity:PlayerEntity = playerEntity.save()
